stage0/cleanrl_framework/requirements.txt
absl-py==2.1.0
albumentations==1.3.1
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docstring_parser==0.16
eval_type_backport==0.2.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.3
grpcio==1.68.1
gymnasium==0.28.1
imageio==2.31.5
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax-jumpy==1.0.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
opencv-python-headless==4.8.1.78
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
protobuf==5.29.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qudida==0.0.4
rich==13.9.4
scikit-image==0.22.0
seaborn==0.12.0
shtab==1.7.1
six==1.17.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tqdm==4.67.1
typeguard==4.4.1
typing_extensions==4.12.2
tyro==0.9.2
tzdata==2024.2
Werkzeug==3.1.3
zipp==3.21.0
